udp_proto.cpp: Bail out of delta encoding when too big to fix msg size overflow
udp_proto.cpp asserts that the number of bits written into
`msg->u.input.bits` is less than MAX_COMPRESSED_BITS to combat buffer
overruns. However, inefficiency in the delta encoding when serializing
inputs with large numbers of changes between frames can overrun that in high
ping situations.

Each bit that changes in the inputs is represented by _10 bits_ on the
wire- a signalling bit, the state of the input, and the position of the
input. Finally, a 0 terminator is written. For human-speed digital inputs,
most frames don't have any change to any bits, so no bits are needed at all.
However, the pathological case is _terrible_: if the session's input is
4 bytes in length and each bit of the input has changed, that frame
requires _320 bits_ when serialized. In high ping situations where a
client may have 13 pending frames, there may be 4160 or more bits
written to `msg->u.input.bits`- resulting in a GGPO failure despite the
fact that the 4-byte input is well under GGPO's limit of
`GAMEINPUT_MAX_BYTES` (9). This obviously gets worse under the following
conditions:

* The session uses an input size closer to the maximum
* Non-digital inputs are used (ex. XInput's 4-bytes per analog stick)
* Non-human tooling is used, like randomizing inputs to look for desyncs

This trades off the zero-bit per frame lower bound with a one-bit per
frame lower bound. That bit is used to determine if delta encoding is
being used. If delta encoding were to result in memory usage greater
than just sending the inputs directly, we now just send the inputs
directly. For the trade, we reduce the upper bound by a factor of 10.